Restored null-date logic.
I find this logic kinda questionable, but it makes some sense.
Basically it is:
If any of the rows (that represent events) found when parsing the list
of events on a press of the "Jump to Date Button" have a null
"data-date" attribute, pretend that event is happening today.

My previous commit changed this logic subtly such that it didn't make
any sense - checking if the result of `dayjs` was null, instead of if
the attribute was null, in 2 of 3 cases - and forgetting to check at all
in the 3rd case (whoops).